+++ app/src/main/res/values/strings.xml
@@ -117,6 +117,7 @@
     <!-- Preferences -->
     <!-- Subsections -->
     <string name="pref_category_general">General</string>
+    <string name="pref_category_library">Library</string>
     <string name="pref_category_reader">Reader</string>
     <string name="pref_category_downloads">Downloads</string>
     <string name="pref_category_sources">Sources</string>
@@ -126,6 +127,17 @@
     <string name="pref_category_about">About</string>
 
     <!-- General section -->
+    <string name="pref_theme">App theme</string>
+    <string name="light_theme">Light</string>
+    <string name="dark_theme">Dark</string>
+    <string name="amoled_theme">AMOLED dark</string>
+    <string name="darkblue_theme">Dark blue</string>
+    <string name="pref_start_screen">Start screen</string>
+    <string name="pref_language">Language</string>
+    <string name="system_default">System default</string>
+    <string name="pref_date_format">Date format</string>
+
+      <!-- Library section -->
     <string name="pref_library_columns">Library manga per row</string>
     <string name="portrait">Portrait</string>
     <string name="landscape">Landscape</string>
@@ -151,15 +163,6 @@
     <string name="charging">Charging</string>
     <string name="pref_update_only_non_completed">Only update ongoing manga</string>
     <string name="pref_auto_update_manga_sync">Sync chapters after reading</string>
-    <string name="pref_theme">Application theme</string>
-    <string name="light_theme">Main theme</string>
-    <string name="dark_theme">Dark theme</string>
-    <string name="amoled_theme">AMOLED theme</string>
-    <string name="darkblue_theme">Dark blue</string>
-    <string name="pref_start_screen">Start screen</string>
-    <string name="pref_language">Language</string>
-    <string name="system_default">System default</string>
-    <string name="pref_date_format">Date format</string>
     <string name="default_category">Default category</string>
     <string name="default_category_summary">Always ask</string>
